Move useHalfTierMods into GeneratedSets
